+++ src/dynamite/_backend/bcuda_impl.h
@@ -1,5 +1,5 @@
 
 #include <petscdevice_cuda.h>
 
-#include "shell_context.h"
 #include "bsubspace_impl.h"
+#include "shell_context.h"