+++ src/dynamite/_backend/bcuda_template_2_private.h
@@ -2,10 +2,7 @@
 #include <cuda_runtime.h>
 #include <thrust/device_ptr.h>
 #include <petscmat.h>
-#include "bcuda_template.h"
-
-#define GPU_BLOCK_SIZE 1024
-#define GPU_BLOCK_NUM 1024
+#include "bcuda_template_2.h"
 
 PetscErrorCode C(BuildContext_CUDA,C(LEFT_SUBSPACE,RIGHT_SUBSPACE))(
   const msc_t *msc,
@@ -26,6 +23,7 @@ __global__ void C(device_MatMult,C(LEFT_SUBSPACE,RIGHT_SUBSPACE))(
   PetscInt nmasks,
   C(data,LEFT_SUBSPACE) *left_subspace_data,
   C(data,RIGHT_SUBSPACE) *right_subspace_data,
+  PetscReal* diag,
   const PetscScalar* xarray,
   PetscScalar* barray);